fix 9711
